Refactor validator and local validator
